Refactored the profile buttons
